requirements.txt
# tkinter
colour-science
pillow
matplotlib
scipy